requirements.txt
langchain
langchain_community
chromadb 
sentence-transformers
google-generativeai
pypdf
langchain-chroma
langchain-huggingface